backend/requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.6,<5
